requirements.txt
attrs==21.4.0
certifi==2022.6.15
charset-normalizer==2.0.12
idna==3.3
iniconfig==1.1.1
oauthlib==3.2.0
packaging==21.3
pluggy==1.0.0
psycopg2-binary==2.9.3
py==1.11.0
pyparsing==3.0.9
pytest==7.1.2
requests==2.28.0
requests-oauthlib==1.3.1
tomli==2.0.1
tweepy==4.10.0
urllib3==1.26.9
